move helper method
